Remove Certguard and File installation as plugin
It was now merged into pulpcore
