--- manifests/plugin/certguard.pp
@@ -1,5 +1,0 @@
-# @summary Pulp Certguard plugin
-class pulpcore::plugin::certguard {
-  pulpcore::plugin { 'certguard':
-  }
-}
