+++ manifests/plugin/deb.pp
@@ -1,14 +1,14 @@
 # @summary Pulp Deb plugin
-# @param use_pulp2_content_route
+# @param use_pulp2_file_content_route
 #   Whether to redirect the legacy (Pulp 2) URL, /pulp/deb/, to the content server
 # @param force_ignore_missing_package_indices
 #   Wheter to set the FORCE_IGNORE_MISSING_PACKAGE_INDICES setting to True or
 #   False in /etc/pulp/settings.py.
 class pulpcore::plugin::deb (
-  Boolean $use_pulp2_content_route = false,
+  Boolean $use_pulp2_file_content_route = false,
   Boolean $force_ignore_missing_package_indices = true,
 ) {
-  if $use_pulp2_content_route {
+  if $use_pulp2_file_content_route {
     $context = {
       'directories' => [
         {
